Update README.md
Big-time Manhattan corporate lawyer Harvey Specter and his team, which includes Donna Paulsen, Louis Litt, and Alex Williams, are launched into a play for power when a new partner joins the firm. With his two best associates gone and Jessica back in Chicago, Specter and the team try to adjust to a new normal without them. The team faces down betrayals, fiery relationships, and secrets that eventually come to light. Old and new rivalries surface among members of the team as they learn to deal with their new member.